\name{qrage-package}
\alias{qrage-package}
\docType{package}
\title{
Tools that Create D3 JavaScript Force Directed Graph from R
}
\description{
Tools that creating D3 JavaScript Force Directed Graph for R. D3 JavaScript was created by Michael Bostock. See http://d3js.org/ and, more specifically for Force Directed Graph https://github.com/mbostock/d3/wiki/Force-Layout. 
}
\details{
\tabular{ll}{
Package: \tab qrage\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-06-21\cr
License: \tab MIT + file LICENSE\cr
}

}
\author{
Shingo Yamamoto

Maintainer: Shingo Yamamoto <yamamototto@gmail.com>
}

\keyword{ 
package 
}
